Library

MultiprecisionExponential.exp_mp — Function

exp_mp(A, kwargs...)

Computes the exponential of $A$ in arbitrary precision, using the algorithm described in [hf19_mpexpm]

Arguments

A::AbstractMatrix: The matrix whose exponential we want to compute

Keyword arguments

working_precision::Integer: The working precision of the algorithm. Defaults to the precision of the type of the elements of $A$

epsilon: tolerance on the error bound used in the parameter selection phase. If set to nothing, epsilon will equal the unit roundoff of the type of the elements of $A$. Defaults to nothing

maxscaling::Integer: the maximum allowed value for the sclaing parameter $s$. Defaults to $100$

maxdegree::Integer: the maximum allowed approximant degree $m$. Defaults to $500$

algorithm::Symbol: the algorithmic variant to use. It must be one of :transfree, :realschur or :complexschur. The :realschur and :complexschur variants compute a (real and complex, respectively) Schur form $A=QXQ^H$; the algorithm is then executed on $X$, thereby computing an approximation $R$ to $\exp(X)$. Finally, $QRQ^H$ is returned. The :transfree variant computes no Schur form. Note that the distinction between :realschur and :complexschur only makes sense if $A$ is real. See remark below. Defaults to :transfree.

approximant::Symbol: The approximant to use. It must be either :taylor or :diagonalcheap, corresponding to Taylor and diagonal Padé approximants, respectively. Defaults to :diagonalcheap

use_abs_err::Bool: whether to use the absolute error in the stopping criterion. Defaults to false.

use_extra_precision::Bool: whether to increase the precision in some internal computations (such as evaluating matrix polynomials) performed when the upper bound is evaluated. The local increase is by a factor of 1.2. Defaults to true.

Remarks

The output matrix will be of type BigFloat (or Complex{BigFloat}, if $A$ is complex)

Note that :realschur and :complexschur have the same behaviour if $A$ is complex, as schur(A) computes the complex Schur form by default in this case. Thus, the distinction only makes sense for real matrices.

Example

julia> using LinearAlgebra

julia> A = [big(π)im 0; 0 big(π)im]
2×2 Matrix{Complex{BigFloat}}:
 0.0+3.14159im  0.0+0.0im
 0.0+0.0im      0.0+3.14159im

julia> Q, _ = qr(randn(BigFloat, 2,2));

julia> Y = exp_mp(Q*A*Q');

julia> norm(Q'*Y*Q + I(2)) ≤ 10eps(BigFloat)
true

References

[hf19_mpexpm] M. Fasi and N. J. Higham, An Arbitrary Precision Scaling and Squaring Algorithm for the Matrix Exponential SIAM J. Matrix Anal. Appl., Vol. 40.4 (2019), pp.1233-1256. doi: 10.1137/18M1228876